fix: extract regex logic into utility function
